Use signed int for reverse path loops to prevent underflow
Change loop index from uint8_t to int in reverse path iteration loops.
When PathsCount is 0, the expression PathsCount - 1 now evaluates to -1
(via integer promotion) instead of wrapping to 255, so the loop condition
'i > 0' is immediately false and the loop body is never entered.

Cast to (uint8_t) at the QuicPathRemove call sites to satisfy warnings-
as-errors for the int-to-uint8_t narrowing conversion.

+++ src/core/connection.c
@@ -5919,17 +5919,16 @@ QuicConnRecvDatagrams(
     // NB: Traversing the array backwards is simpler and more efficient here due
     // to the array shifting that happens in QuicPathRemove.
     //
-    for (uint8_t i = Connection->PathsCount - 1; i > 0; --i) {
+    for (int i = Connection->PathsCount - 1; i > 0; --i) {
         if (!Connection->Paths[i].GotValidPacket) {
             QuicTraceLogConnInfo(
                 PathDiscarded,
                 Connection,
                 "Removing invalid path[%hhu]",
                 Connection->Paths[i].ID);
-            QuicPathRemove(Connection, i);
+            QuicPathRemove(Connection, (uint8_t)i);
         }
     }
-
     if (!Connection->State.UpdateWorker && Connection->State.Connected &&
         !Connection->State.ShutdownComplete && RecvState.UpdatePartitionId) {
         //
